requirements.txt
crowdstrike-falconpy>=1.4.0

# テスト用依存パッケージ
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
